[Will Branch] - Implemented water hand interaction
- Added god rays
