+++ Assets/Packages/HandAssets/HandManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b0ce7ddb8739f2940ae66d7357b95f00
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
