+++ Assets/Packages/HandAssets/Prefabs/Hands.prefab
@@ -1,113 +1,5 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &895096352
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 895096356}
-  - component: {fileID: 895096355}
-  - component: {fileID: 895096354}
-  - component: {fileID: 895096353}
-  - component: {fileID: 895096357}
-  m_Layer: 0
-  m_Name: WaterSplashTestCube
-  m_TagString: Player
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 0
---- !u!4 &895096356
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 895096352}
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -10.2604065, y: 3.73, z: 6.25}
-  m_LocalScale: {x: 0.5, y: 0.5, z: 0.5}
-  m_Children: []
-  m_Father: {fileID: 4706324235734989328}
-  m_RootOrder: 2
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!33 &895096355
-MeshFilter:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 895096352}
-  m_Mesh: {fileID: 10202, guid: 0000000000000000e000000000000000, type: 0}
---- !u!23 &895096354
-MeshRenderer:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 895096352}
-  m_Enabled: 1
-  m_CastShadows: 0
-  m_ReceiveShadows: 0
-  m_DynamicOccludee: 1
-  m_MotionVectors: 1
-  m_LightProbeUsage: 0
-  m_ReflectionProbeUsage: 0
-  m_RenderingLayerMask: 1
-  m_RendererPriority: 0
-  m_Materials:
-  - {fileID: 10303, guid: 0000000000000000f000000000000000, type: 0}
-  m_StaticBatchInfo:
-    firstSubMesh: 0
-    subMeshCount: 0
-  m_StaticBatchRoot: {fileID: 0}
-  m_ProbeAnchor: {fileID: 0}
-  m_LightProbeVolumeOverride: {fileID: 0}
-  m_ScaleInLightmap: 1
-  m_PreserveUVs: 0
-  m_IgnoreNormalsForChartDetection: 0
-  m_ImportantGI: 0
-  m_StitchLightmapSeams: 1
-  m_SelectedEditorRenderState: 3
-  m_MinimumChartSize: 4
-  m_AutoUVMaxDistance: 0.5
-  m_AutoUVMaxAngle: 89
-  m_LightmapParameters: {fileID: 0}
-  m_SortingLayerID: 0
-  m_SortingLayer: 0
-  m_SortingOrder: 0
---- !u!65 &895096353
-BoxCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 895096352}
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_Enabled: 1
-  serializedVersion: 2
-  m_Size: {x: 1, y: 1, z: 1}
-  m_Center: {x: 0, y: 0, z: 0}
---- !u!54 &895096357
-Rigidbody:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 895096352}
-  serializedVersion: 2
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: 0.05
-  m_UseGravity: 0
-  m_IsKinematic: 0
-  m_Interpolate: 0
-  m_Constraints: 0
-  m_CollisionDetection: 0
 --- !u!1 &4706324235734989329
 GameObject:
   m_ObjectHideFlags: 0
@@ -117,6 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 4706324235734989328}
+  - component: {fileID: 4893542585136034988}
   m_Layer: 0
   m_Name: Hands
   m_TagString: Untagged
@@ -137,10 +30,23 @@ Transform:
   m_Children:
   - {fileID: 4967656505474361987}
   - {fileID: 5066438307026241592}
-  - {fileID: 895096356}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &4893542585136034988
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4706324235734989329}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b0ce7ddb8739f2940ae66d7357b95f00, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _rightHand: {fileID: 8262962500344207511}
+  _leftHand: {fileID: 8076294526226790956}
 --- !u!1001 &4706324237372906989
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -215,6 +121,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3550239352464451521, guid: 9d6dbace3344a91479b8b43560a3f2f3,
         type: 3}
+      propertyPath: _handManager
+      value: 
+      objectReference: {fileID: 4893542585136034988}
+    - target: {fileID: 3550239352464451521, guid: 9d6dbace3344a91479b8b43560a3f2f3,
+        type: 3}
       propertyPath: _isRightHand
       value: 0
       objectReference: {fileID: 0}
@@ -226,6 +137,18 @@ Transform:
     type: 3}
   m_PrefabInstance: {fileID: 4706324237372906989}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &8076294526226790956 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 3550239352464451521, guid: 9d6dbace3344a91479b8b43560a3f2f3,
+    type: 3}
+  m_PrefabInstance: {fileID: 4706324237372906989}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 321549c20ad6a5c49b2bfb1bf1c4e50a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &4895143951442542422
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -293,6 +216,11 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 270
       objectReference: {fileID: 0}
+    - target: {fileID: 3550239352464451521, guid: 9d6dbace3344a91479b8b43560a3f2f3,
+        type: 3}
+      propertyPath: _handManager
+      value: 
+      objectReference: {fileID: 4893542585136034988}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 9d6dbace3344a91479b8b43560a3f2f3, type: 3}
 --- !u!4 &4967656505474361987 stripped
@@ -301,3 +229,15 @@ Transform:
     type: 3}
   m_PrefabInstance: {fileID: 4895143951442542422}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &8262962500344207511 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 3550239352464451521, guid: 9d6dbace3344a91479b8b43560a3f2f3,
+    type: 3}
+  m_PrefabInstance: {fileID: 4895143951442542422}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 321549c20ad6a5c49b2bfb1bf1c4e50a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
